Added tag autocomplete mode and some minor changes
Setting the mode option to tag causes the function to complete basing upon the characters inserted after the last comma.
The suggestion box is inserted now in the document.body to avoid offset calculations which depends on position style value (relative or not).
The element may be passed as a string id or an element itself (may be useful to not assign to much ids in the html document).
Some demo fixes and new tag example provided.